Added authorization to device endpoints

+++ backend/Controllers/DevicesController.cs
@@ -15,26 +15,18 @@ public class DevicesController : ControllerBase
         this._dbContext = dataContext;
     }
 
+    [Authorize]
     [HttpGet]
     [Route("/devices")]
     public async Task<ActionResult<List<Device>>> Get()
     {
-        var devices = new List<Device> {
-       new Device("Light bulb", "RGB light buleqwedqwdbqwdbqwbdkqhwgdkqhwkdhkqwjhdkjqwhkjdhb"),
-       new Device("Air conditioner"),
-       new Device("Light bulb", "RGB light bulb"),
-       new Device("Air conditioner"),
-       new Device("Light bulb", "RGB light bulb"),
-       new Device("Air conditioner"),
-       new Device("Light bulb", "RGB light bulb"),
-       new Device("Air conditioner")
-     };
-
-        //var devices = _dbContext.Devices.ToList();
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        var devices = _dbContext.Devices.Where(d => d.UserId == Int32.Parse(userId)).ToList();
 
         return Ok(devices);
     }
 
+    [Authorize]
     [HttpGet]
     [Route("{deviceId}")]
     public async Task<ActionResult<Device>> GetById(string deviceId)
@@ -50,16 +42,19 @@ public class DevicesController : ControllerBase
         {
             return NotFound("Device not found");
         }
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if (device.UserId != Int32.Parse(userId))
+        {
+            return Unauthorized();
+        }
 
         return Ok(device);
     }
 
-    [HttpPost, Authorize]
+    [Authorize]
+    [HttpPost]
     public async Task<ActionResult<Device>> addDevice(AddDeviceDto deviceData)
     {
-        // device.Id = Guid.NewGuid();
-        // this._dbContext.Devices.Add(device);
-        // await _dbContext.SaveChangesAsync();
         var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
         System.Console.WriteLine(userId);
         using (var httpClient = new HttpClient())
@@ -71,6 +66,7 @@ public class DevicesController : ControllerBase
             device.Href = data.Href;
             device.Properties = data.Properties;
             device.Actions = data.Actions;
+            device.UserId = Int32.Parse(userId);
             this._dbContext.Devices.Add(device);
             await _dbContext.SaveChangesAsync();
             return Ok(device);
@@ -85,6 +81,7 @@ public class DevicesController : ControllerBase
         return Ok(dbDevice);
     }
 
+    [Authorize]
     [HttpPut]
     [Route("{deviceId}")]
     public async Task<ActionResult<Device>> updateDevice(string deviceId, Device device)
@@ -108,6 +105,7 @@ public class DevicesController : ControllerBase
         return Ok(device);
     }
 
+    [Authorize]
     [HttpDelete]
     [Route("{deviceId}")]
     public async Task<ActionResult> deleteDevice(string deviceId)
@@ -123,12 +121,19 @@ public class DevicesController : ControllerBase
         {
             return NotFound("Device not found");
         }
+        
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if (dbDevice.UserId != Int32.Parse(userId))
+        {
+            return Unauthorized();
+        }
 
         this._dbContext.Devices.Remove(dbDevice);
         await this._dbContext.SaveChangesAsync();
         return Ok(true);
     }
 
+    [Authorize]
     [HttpGet]
     [Route("{deviceId}/properties")]
     public async Task<ActionResult> GetProperties(string deviceId)
@@ -145,6 +150,12 @@ public class DevicesController : ControllerBase
             return NotFound("Device not found");
         }
 
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if (device.UserId != Int32.Parse(userId))
+        {
+            return Unauthorized();
+        }
+
         using (var httpClient = new HttpClient())
         {
             var data = await httpClient.GetFromJsonAsync<Dictionary<string, object>>(new Uri($"{device.Base}/properties"));
@@ -153,7 +164,7 @@ public class DevicesController : ControllerBase
         }
     }
 
-
+    [Authorize]
     [HttpPatch]
     [Route("{deviceId}/properties/{propertyName}")]
     public async Task<ActionResult> updateProperty(string deviceId, string propertyName, [FromBody]object newVal)
@@ -170,15 +181,21 @@ public class DevicesController : ControllerBase
             return NotFound("Device not found");
         }
 
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if (device.UserId != Int32.Parse(userId))
+        {
+            return Unauthorized();
+        }
+
         using (var httpClient = new HttpClient())
         {
-
             var data = await httpClient.PutAsJsonAsync(new Uri($"{device.Base}/properties/{propertyName}"), new Dictionary<string, object> {{propertyName, newVal}});
             System.Console.WriteLine(data);
             return Ok(data);
         }
     }
 
+    [Authorize]
     [HttpPut]
     [Route("{deviceId}/{command}")]
     public async Task<ActionResult> executeAction(string deviceId, string command, [FromBody] string newVal)
